requirements.txt
djangorestframework
djongo
django-cors-headers